+++ elpee/algorithms/dual_simplex.py
@@ -1,5 +1,5 @@
 from typing import List
-from elpee.utils.protocols.LPProblem import LPProblem
+from elpee.utils.protocols.lp_problem import LPProblem
 
 M = 1000000
